Integrate external netcdf with apply_oe (#439)
- removes multisurface_oe (to be re-integrated later)
- migrates core apply_oe functions to template_construction
- sets up pass-throughs to allow for an external netcdf to be referenced in apply_oe
- re-organizes apply_oe
- applies some patch fixes to lut.py
