+++ isofit/radiative_transfer/luts.py
@@ -205,9 +205,11 @@ def sel(ds, dim, lt=None, lte=None, gt=None, gte=None, encompass=True):
     # Retrieve the previous/next values such that gte and lte are encompassed
     if encompass:
         if gte is not None:
-            gte = ds[dim].where(ds[dim] < gte).dropna(dim)[-1]
+            loc_subset = ds[dim].where(ds[dim] < gte).dropna(dim)
+            gte = loc_subset[-1] if len(loc_subset) > 0 else ds[dim].min()
         if lte is not None:
-            lte = ds[dim].where(ds[dim] > lte).dropna(dim)[0]
+            loc_subset = ds[dim].where(ds[dim] > lte).dropna(dim)
+            lte = loc_subset[0] if len(loc_subset) > 0 else ds[dim].max()
 
     if lt is not None:
         ds = ds.sel({dim: ds[dim] < lt})
