+++ isofit/utils/analytical_line.py
@@ -17,6 +17,7 @@
 # ISOFIT: Imaging Spectrometer Optimal FITting
 # Author: Philip G. Brodrick, philip.brodrick@jpl.nasa.gov
 
+import atexit
 import logging
 import multiprocessing
 import os
@@ -30,7 +31,7 @@ from spectral.io import envi
 
 from isofit import ray
 from isofit.configs import configs
-from isofit.core.common import envi_header, load_spectrum
+from isofit.core.common import envi_header, load_spectrum, ray_initiate
 from isofit.core.fileio import write_bil_chunk
 from isofit.core.forward import ForwardModel
 from isofit.core.geometry import Geometry
@@ -66,7 +67,7 @@ def analytical_line(
     )
 
     if isofit_config is None:
-        file = glob(os.path.join(isofit_dir, "config", "") + "*_modtran.json")[0]
+        file = glob(os.path.join(isofit_dir, "config", "") + "*_isofit.json")[0]
     else:
         file = isofit_config
 
@@ -144,13 +145,17 @@ def analytical_line(
     )
     del atm, rdn, img
 
-    ray.init(
-        ignore_reinit_error=config.implementation.ray_ignore_reinit_error,
-        address=config.implementation.ip_head,
-        _temp_dir=config.implementation.ray_temp_dir,
-        include_dashboard=config.implementation.ray_include_dashboard,
-        _redis_password=config.implementation.redis_password,
-    )
+    ray_dict = {
+         "ignore_reinit_error": config.implementation.ray_ignore_reinit_error,
+         "address": config.implementation.ip_head,
+         "_temp_dir": config.implementation.ray_temp_dir,
+         "include_dashboard": config.implementation.ray_include_dashboard,
+         "_redis_password": config.implementation.redis_password,
+         "num_cpus": n_cores,
+    }
+
+    ray_initiate(ray_dict)
+    atexit.register(ray.shutdown)
 
     n_workers = n_cores
     if n_workers == -1:
@@ -170,7 +175,7 @@ def analytical_line(
     ]
     workers = ray.util.ActorPool([worker.remote(*wargs) for _ in range(n_workers)])
 
-    line_breaks = np.linspace(0, rdns[0], n_workers * 3, dtype=int)
+    line_breaks = np.linspace(0, rdns[0], n_workers * config.implementation.task_inflation_factor, dtype=int)
     line_breaks = [
         (line_breaks[n], line_breaks[n + 1]) for n in range(len(line_breaks) - 1)
     ]
@@ -183,6 +188,7 @@ def analytical_line(
         f"{round(rdns[0]*rdns[1]/total_time,4)} spectra/s, "
         f"{round(rdns[0]*rdns[1]/total_time/n_workers,4)} spectra/s/core"
     )
+    ray.shutdown()
 
 
 class Worker(object):
